refactored SecureRpcHelper into base class
